Textarea shows acronym and short-form with label
Typeahead text-area no longer shows only the label of each suggestion
set entry when selected but also the acronym of its ontology of origin
and short-form (i.e. term id) similar to previous improvement applied
to entries of the suggestion set selectable list.

This required that a function combining content of multiple fields in a
single string be passed as “display” parameter.